requirements.txt
asgiref==3.11.0
certifi==2026.1.4
channels==4.3.2
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysql-connector-python==9.5.0
mysqlclient==2.2.7
numpy==2.4.0
package_name==0.1
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
